requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.3
pydantic-settings==2.3.3

# Computer vision
mediapipe==0.10.14
opencv-python-headless==4.9.0.80
numpy==1.26.4

# Signal processing
scipy>=1.13.0

# Dashboard
gradio>=5.9.0

# AWS SDK
boto3==1.34.131
botocore==1.34.131

# Testing
pytest>=9.0.3
httpx==0.27.0
pytest-asyncio==0.23.7

# Security: pin transitive deps with known CVEs
pillow>=12.2.0

# ML Pipeline & Experiment Tracking
scikit-learn>=1.4.0
joblib>=1.4.0
mlflow>=2.12.0

# Monitoring & Observability
prometheus-client>=0.20.0

# Utilities
python-dotenv>=1.0.0
